+++ internal/app/types/modes.go
@@ -1,0 +1,8 @@
+package types
+
+const (
+	AddMode    string = "ADD"
+	DeleteMode string = "DELETE"
+	EditMode   string = "EDIT"
+	ReadMode   string = "READ"
+)