+++ internal/app/types/stores.go
@@ -21,3 +21,13 @@ type UsersStore interface {
 	GetByEmail(eml string) (*User, error)
 	Insert(usr *User) (*User, error)
 }
+
+type WebAppsStore interface {
+	dbStore
+	CountByFilter(ftr Filter) (int, error)
+	DeleteByIdAndUserEmail(id, usr string) error
+	GetByFilter(ftr Filter, pag Pagination) ([]*WebApp, error)
+	GetByIdAndUserEmail(id, usr string) (*WebApp, error)
+	Insert(wap *WebApp) (*WebApp, error)
+	Update(wap *WebApp) (*WebApp, error)
+}